Added option to remove previous poylgon. Relates to #27

+++ components/FreeDraw.js
@@ -679,9 +679,19 @@
 
             if (!this.options.multiplePolygons && this.getPolygons(true).length >= 1) {
 
-                // User is only allowed to create one polygon.
-                this.destroyD3().createD3();
-                return false;
+                if (this.options.destroyPrevious) {
+
+                    // Destroy the current polygon and then draw the current polygon.
+                    this.silently(this.clearPolygons);
+
+                } else {
+
+                    // Otherwise delete the line because polygon creation has been disallowed, since there's
+                    // already one polygon on the map.
+                    this.destroyD3().createD3();
+                    return false;
+
+                }
 
             }
